+++ skills/interaction/SKILL.md
@@ -6,21 +6,3 @@
 - 仅在必要时使用，避免频繁打断用户
 - 问题应清晰明确，说明需要用户提供什么信息
 - CLI 模式下通过 stdin 获取输入，QQ Bot 模式下通过消息回复获取
-
-## call_sub_agent
-创建一个独立的子 Agent 会话，将任务委派给子 Agent 执行单轮对话，子 Agent 的最终回答会作为工具返回值。
-
-**使用场景**：
-- 需要独立上下文窗口的复杂分析任务（如分析大文件）
-- 多步骤搜索 + 汇总，需要子 Agent 独立进行工具调用
-- 并行调用多个子 Agent 处理不同子问题
-
-**参数**：
-- `task`：子 Agent 的用户提示词（完整任务描述）
-- `name`：可选，子会话的显示名称（用于侧边栏标识）
-
-**注意事项**：
-- 子 Agent 拥有独立上下文，不会污染主对话
-- 子 Agent 只能执行一轮对话（单 turn）
-- 子会话在前端侧边栏中以只读模式展示
-- 嵌套深度限制为 2 层
